Remove deprecated interfaces and update QueryInterface
